Added TriggerResults, compile and run checked. Added also Event Header info but still missing run check.
Ignore-this: 55672439bacd39f336409b4e2c236ab7
Note: for the trigger results, it could be redefined the MTAtomBuilder in order to accept the ParameterSet
passed by the MiniTreesProducer constructor (i.e. the configuration python file). In this way, we can
include another key in the config python to select the trigger paths we want to evaluate. This approach
could be dangerous in the sense we can't forget to put some trigger path in some processing and if we
don't have a good system to store what was used in a processing we could introduce mistakes difficult
to find. If we hardcode the trigger paths inside the class, by the other way, we are assuring the trace
of what paths was used in a determined processing just watching what version of the MiniTreesProducer
was used (using for that the cvs, svn or whatever system of repositories we have). Of course, we can
include the config python in the repository...

darcs-hash:20110304091104-e973b-7b34070aa74a2228a445703d1eaf67b713061fa9.gz