Added Muons (pat::Muon). Still to include pfMuons which share a lot Muon's functionalities (maybe a common class Lepton??)
Ignore-this: cb3974357c0b8b82c4663f3c13d1b0bd

Some problems with the current implementation of the class. Muons need the vertex information, but as the classes
are implemented there is no way to access the vertex information (as the instance label,...) and it is inefficient
in the sense we have to access twice for the vertex collection or others such beam spot, etc ... It seems necessary
to add another class to deal with this kind of shared information, collections which will be used by more than one
object. Looking for template... aggregation relationship??

darcs-hash:20110304164640-e973b-55c68134f97aec76cad108b864b9c31411c5d3be.gz